+++ CHANGELOG.md
@@ -6,6 +6,11 @@ This changelog is following the recommended format by [keepachangelog](https://k
 
 ## [Unreleased]
 
+### Added
+
+- Add support for Search attribute config (cf. [#64](https://github.com/yannick-beot-sp/vscode-sailpoint-identitynow/issues/64))
+- Add support for Identity Attributes
+
 ## [0.0.29] - 2024-02-28
 
 ### Added